Remove the legacy PasteDeploy code.
Full PasteDeploy compatibility will be in an add-on; it was getting too complicated.
