--- docs/reference/montague.compat.config.rst
@@ -1,5 +1,0 @@
-montague.compat.config
-=============================
-
-.. automodule:: montague.compat.config
-    :members:
