--- docs/reference/montague.compat.converters.rst
@@ -1,6 +1,0 @@
-montague.compat.converters
-=============================
-
-.. automodule:: montague.compat.converters
-    :members:
-    :undoc-members:
