--- docs/reference/montague.compat.loadwsgi.rst
@@ -1,6 +1,0 @@
-montague.compat.loadwsgi
-=============================
-
-.. automodule:: montague.compat.loadwsgi
-    :members:
-    :undoc-members:
