--- docs/reference/montague.compat.rst
@@ -1,6 +1,0 @@
-montague.compat
-=============================
-
-.. automodule:: montague.compat
-    :members:
-    :undoc-members:
