--- docs/reference/montague.compat.util.rst
@@ -1,5 +1,0 @@
-montague.compat.util
-=============================
-
-.. automodule:: montague.compat.util
-    :members:
