+++ src/montague/__init__.py
@@ -1,8 +1,8 @@
 from __future__ import absolute_import
 
-__version__ = "0.1.0"
-
 from .loadwsgi import Loader
+
+__version__ = "0.1.0"
 
 
 def load_app(config_path, name=None):
