--- src/montague/compat/__init__.py
@@ -1,6 +1,0 @@
-from __future__ import absolute_import
-# (c) 2005 Ian Bicking and contributors; written for Paste (http://pythonpaste.org)
-# Licensed under the MIT license: http://www.opensource.org/licenses/mit-license.php
-from .loadwsgi import loadapp, loadserver, loadfilter, appconfig
-
-__all__ = ['loadapp', 'loadserver', 'loadfilter', 'appconfig']
