--- tests/compat/sample_configs/basic_app.ini
@@ -1,14 +1,0 @@
-[application:main]
-use = egg:FakeApp#basic_app
-
-[application:other]
-use = egg:FakeApp#other
-
-[composit:remote_addr]
-use = egg:FakeApp#remote_addr
-app.1 = main
-addr.1 = 127.0.0.1
-
-app.2 = other
-addr.2 = 0.0.0.0
-
