--- tests/compat/sample_configs/executable.ini
@@ -1,10 +1,0 @@
-#!/usr/bin/env paster
-[exe]
-sys.path = /fake/path/
-  /another/fake/path ../fake_packages/
-
-[server]
-use = egg:PasteScript#cgi
-
-[app]
-use = egg:FakeApp#basic_app
