--- tests/compat/sample_configs/test_config.ini
@@ -1,38 +1,0 @@
-[DEFAULT]
-def1 = a
-def2 = b
-basepath = %(here)s
-
-[app:test1]
-use = egg:FakeApp#configed
-setting1 = foo
-setting2 = bar
-apppath = %(basepath)s/app
-
-[app:test2]
-use = egg:FakeApp#configed
-set def1 = test2
-set another = TEST
-local conf = something
-
-[app:test3]
-use = test2
-set def1 = test3
-another = something more  
-    across several
-    lines
-
-[app:test_foreign_config]
-use = config:test_config_included.ini
-set glob = override
-another = FOO
-
-[app:test_get]
-use = egg:FakeApp#configed
-set def2 = TEST
-get def1 = def1
-get foo = def2
-
-[app:test_global_conf]
-use = egg:FakeApp#configed
-test_interp = this:%(inherit)s
