--- tests/compat/sample_configs/test_config_included.ini
@@ -1,10 +1,0 @@
-[DEFAULT]
-def2 = from include
-def3 = c
-
-[app:main]
-# Equivalent to the egg reference, but just for kicks...
-paste.app_factory = fakeapp.configapps:SimpleApp.make_app
-set glob = orig
-bob = your uncle
-another = BAR
