--- tests/compat/sample_configs/test_error.ini
@@ -1,8 +1,0 @@
-[DEFAULT]
-def1 = a
-def2 = b
-
-[app:main]
-use = egg:FakeApp#configed
-setting1 = foo
-setting2 = %(does_not_exist)s/bar
