--- tests/compat/sample_configs/test_filter.ini
@@ -1,22 +1,0 @@
-[app:normal]
-use = egg:FakeApp#basic_app
-
-[pipeline:piped]
-pipeline = egg:FakeApp#caps normal
-
-[filter-app:filt]
-use = egg:FakeApp#caps
-method_to_call = lower
-next = normal
-
-[pipeline:piped2]
-pipeline = egg:FakeApp#caps2 normal
-
-[filter-app:filt2]
-use = egg:FakeApp#caps2
-method_to_call = lower
-next = normal
-
-[app:inv]
-use = egg:FakeApp#basic_app
-filter-with = egg:FakeApp#caps
