--- tests/compat/sample_configs/test_filter_with.ini
@@ -1,12 +1,0 @@
-[app:main]
-use = egg:FakeApp#basic_app
-example = test
-filter-with = filter1
-
-[filter:filter1]
-use = egg:FakeApp#caps
-filter-with = filter2
-
-[filter:filter2]
-use = egg:FakeApp#caps
-
