--- tests/compat/sample_configs/test_func.ini
@@ -1,13 +1,0 @@
-[application:main]
-use = call:fakeapp.apps:make_basic_app
-
-[application:other]
-use = call:fakeapp.apps:make_basic_app2
-
-[composit:remote_addr]
-use = call:fakeapp.apps:make_remote_addr
-app.1 = main
-addr.1 = 127.0.0.1
-
-app.2 = other
-addr.2 = 0.0.0.0
