--- tests/compat/test_converters.py
@@ -1,18 +1,0 @@
-def test_asbool_truthy():
-    from montague.compat.converters import asbool
-    assert asbool('true')
-    assert asbool('yes')
-    assert asbool('on')
-    assert asbool('y')
-    assert asbool('t')
-    assert asbool('1')
-
-    
-def test_asbool_falsy():
-    from montague.compat.converters import asbool
-    assert not asbool('false')
-    assert not asbool('no')
-    assert not asbool('off')
-    assert not asbool('n')
-    assert not asbool('f')
-    assert not asbool('0')
